style: `MemoryTypes.cs`改名为`MemoryEntry.cs`并转移其中部分内容
- 部分内容转移至MemoryCategory.cs
